requirements.txt
langchain
langchain-openai
langchain-community
tiktoken
docarray
openai
pydantic
chromadb
python-dotenv
unstructured
pypdf
langchain_chroma
spacy 
python-docx
PyPDF2
flask
unstructured[pdf]

